+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release 5.5.1
  * Fix DCQL Query extraction for SD-JWT
+ * Support more credential identifiers for presentation exchange
  * Update to VC-K 5.5.1
 
 # Release 5.5.0